+++ packages/admin/admin/src/table/TableQuery.tsx
@@ -8,12 +8,18 @@ import { Loading } from "../common/Loading";
 import { styles, TableQueryClassKey } from "./TableQuery.styles";
 import { ITableQueryApi, TableQueryContext } from "./TableQueryContext";
 
+/**
+ * @deprecated Use MUI X Data Grid in combination with `useDataGridRemote` instead.
+ */
 export const parseIdFromIri = (iri: string) => {
     const m = iri.match(/\/(\d+)/);
     if (!m) return null;
     return m[1];
 };
 
+/**
+ * @deprecated Use MUI X Data Grid in combination with `useDataGridRemote` instead.
+ */
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface IDefaultVariables {}
 
@@ -57,6 +63,9 @@ export function Query({ classes, ...otherProps }: TableQueryProps & WithStyles<t
     );
 }
 
+/**
+ * @deprecated Use MUI X Data Grid in combination with `useDataGridRemote` instead.
+ */
 export const TableQuery = withStyles(styles, { name: "CometAdminTableQuery" })(Query);
 
 declare module "@mui/material/styles" {