+++ packages/admin/cms-admin/src/dam/DataGrid/FolderDataGrid.tsx
@@ -335,6 +335,10 @@ const FolderDataGrid = ({
             return "CometDataGridRow--highlighted";
         }
 
+        if (row.isInboxFromOtherScope) {
+            return "CometDataGridRow--inboxFolder";
+        }
+
         return "";
     };
 
@@ -440,7 +444,7 @@ const FolderDataGrid = ({
     ];
 
     return (
-        <div style={{ padding: "20px" }}>
+        <sc.FolderWrapper>
             <FolderHead
                 isSearching={isSearching}
                 numberItems={dataGridData?.damItemsList.totalCount ?? 0}
@@ -490,7 +494,7 @@ const FolderDataGrid = ({
                     closeMoveDialog();
                 }}
             />
-        </div>
+        </sc.FolderWrapper>
     );
 };