+++ packages/api/cms-api/src/mikro-orm/mikro-orm.module.ts
@@ -16,6 +16,8 @@ import { Migration20230209111818 } from "./migrations/Migration20230209111818";
 import { Migration20230302145445 } from "./migrations/Migration20230302145445";
 import { Migration20230613150332 } from "./migrations/Migration20230613150332";
 import { Migration20230802124224 } from "./migrations/Migration20230802124224";
+import { Migration20230821090303 } from "./migrations/Migration20230821090303";
+import { Migration20230831110518 } from "./migrations/Migration20230831110518";
 
 export const PG_UNIQUE_CONSTRAINT_VIOLATION = "23505";
 
@@ -69,6 +71,8 @@ export function createOrmConfig({ migrations, ...defaults }: MikroOrmNestjsOptio
                 { name: "Migration20230209111818", class: Migration20230209111818 },
                 { name: "Migration20230613150332", class: Migration20230613150332 },
                 { name: "Migration20230802124224", class: Migration20230802124224 },
+                { name: "Migration20230821090303", class: Migration20230821090303 },
+                { name: "Migration20230831110518", class: Migration20230831110518 },
                 ...(migrations?.migrationsList || []),
             ],
         },